Feature/chat

+++ src/D2NG/BNCS/Packet/AuthInfoRequestPacket.cs
@@ -4,7 +4,6 @@ using System.Text;
 
 namespace D2NG.BNCS.Packet
 {
-
     public class AuthInfoRequestPacket : BncsPacket
     {
         private static readonly byte[] ProtocolId = BitConverter.GetBytes(0x00);
@@ -35,7 +34,7 @@ namespace D2NG.BNCS.Packet
         public AuthInfoRequestPacket(int version)
             : base(
                 BuildPacket(
-                    (byte)Sid.AUTH_INFO,
+                    Sid.AUTH_INFO,
                     ProtocolId,
                     Encoding.ASCII.GetBytes(PlatformCode).Reverse().ToArray(),
                     Encoding.ASCII.GetBytes(ProductCode).Reverse().ToArray(),
